+++ src/ChannelEffects.jl
@@ -12,25 +12,6 @@ using ..TelemetryCore
 using Dates: Dates, DateTime, Millisecond
 using Random: Random, Xoshiro
 
-export LossModel,
-    NoLoss,
-    BernoulliLoss,
-    GilbertElliottLoss,
-    sample_loss!,
-    stationary_loss_rate,
-    DisruptionEvent,
-    DisruptionTimeline,
-    disruption_factor,
-    disruption_loss_multiplier,
-    active_disruption_label,
-    LinkModel,
-    effective_bandwidth,
-    is_transmittable,
-    build_loss_model,
-    build_disruption_timeline,
-    build_link_model,
-    loss_retry_limit
-
 # --- Stochastic Packet-Loss Models ---
 """
     LossModel